taskflow/backend/requirements.txt
Flask==2.3.3
flask-pymongo==2.3.0
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
flask-cors==4.0.0
pymongo==4.4.0
